+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # opencred-platform Changelog
 
+## 9.0.2 - 2025-06-XX
+
+### Fixed
+- Fixed incorrect status check logic to ensure that revoked credentials.
+- Additional credential correctness and fitness for purpose checks.
+
 ## 9.0.1 - 2025-06-10
 
 ### Fixed